+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symboltable/MethodScope.java
@@ -3,8 +3,10 @@
  */
 package net.sourceforge.pmd.lang.java.symboltable;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.java.ast.ASTConstructorDeclaration;
@@ -27,17 +29,19 @@ public class MethodScope extends AbstractJavaScope {
         return getDeclarations(VariableNameDeclaration.class);
     }
 
-    public NameDeclaration addNameOccurrence(NameOccurrence occurrence) {
+    public Set<NameDeclaration> addNameOccurrence(NameOccurrence occurrence) {
         JavaNameOccurrence javaOccurrence = (JavaNameOccurrence)occurrence;
-        NameDeclaration decl = findVariableHere(javaOccurrence);
-        if (decl != null && !javaOccurrence.isThisOrSuper()) {
-            getVariableDeclarations().get(decl).add(javaOccurrence);
-            Node n = javaOccurrence.getLocation();
-            if (n instanceof ASTName) {
-                ((ASTName) n).setNameDeclaration(decl);
-            } // TODO what to do with PrimarySuffix case?
+        Set<NameDeclaration> declarations = findVariableHere(javaOccurrence);
+        if (!declarations.isEmpty() && !javaOccurrence.isThisOrSuper()) {
+            for (NameDeclaration decl : declarations) {
+                getVariableDeclarations().get(decl).add(javaOccurrence);
+                Node n = javaOccurrence.getLocation();
+                if (n instanceof ASTName) {
+                    ((ASTName) n).setNameDeclaration(decl);
+                } // TODO what to do with PrimarySuffix case?
+            }
         }
-        return decl;
+        return declarations;
     }
 
     public void addDeclaration(NameDeclaration variableDecl) {
@@ -47,13 +51,17 @@ public class MethodScope extends AbstractJavaScope {
         super.addDeclaration(variableDecl);
     }
 
-    public NameDeclaration findVariableHere(JavaNameOccurrence occurrence) {
+    public Set<NameDeclaration> findVariableHere(JavaNameOccurrence occurrence) {
+        Set<NameDeclaration> result = new HashSet<NameDeclaration>();
         if (occurrence.isThisOrSuper() || occurrence.isMethodOrConstructorInvocation()) {
-            return null;
+            return result;
         }
         ImageFinderFunction finder = new ImageFinderFunction(occurrence.getImage());
         Applier.apply(finder, getVariableDeclarations().keySet().iterator());
-        return finder.getDecl();
+        if (finder.getDecl() != null) {
+            result.add(finder.getDecl());
+        }
+        return result;
     }
 
     public String getName() {